completed the task

+++ restaurants/views.py
@@ -3,6 +3,7 @@ from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
 
+
 def signup(request):
     form = SignupForm()
     if request.method == 'POST':
@@ -16,9 +17,10 @@ def signup(request):
             login(request, user)
             return redirect("restaurant-list")
     context = {
-        "form":form,
+        "form": form,
     }
     return render(request, 'signup.html', context)
+
 
 def signin(request):
     form = SigninForm()
@@ -34,17 +36,19 @@ def signin(request):
                 login(request, auth_user)
                 return redirect('restaurant-list')
     context = {
-        "form":form
+        "form": form
     }
     return render(request, 'signin.html', context)
+
 
 def signout(request):
     logout(request)
     return redirect("signin")
 
+
 def restaurant_list(request):
     context = {
-        "restaurants":Restaurant.objects.all()
+        "restaurants": Restaurant.objects.all()
     }
     return render(request, 'list.html', context)
 
@@ -58,7 +62,10 @@ def restaurant_detail(request, restaurant_id):
     }
     return render(request, 'detail.html', context)
 
+
 def restaurant_create(request):
+    if not request.user.is_authenticated:
+        return redirect("signin")
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -68,13 +75,16 @@ def restaurant_create(request):
             restaurant.save()
             return redirect('restaurant-list')
     context = {
-        "form":form,
+        "form": form,
     }
     return render(request, 'create.html', context)
+
 
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not(request.user.is_staff or request.user == restaurant.owner):
+        return redirect("no-access")
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -83,13 +93,16 @@ def item_create(request, restaurant_id):
             item.save()
             return redirect('restaurant-detail', restaurant_id)
     context = {
-        "form":form,
+        "form": form,
         "restaurant": restaurant,
     }
     return render(request, 'item_create.html', context)
 
+
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not(request.user.is_staff or request.user == restaurant_obj.owner):
+        return redirect("no-access")
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -98,11 +111,17 @@ def restaurant_update(request, restaurant_id):
             return redirect('restaurant-list')
     context = {
         "restaurant_obj": restaurant_obj,
-        "form":form,
+        "form": form,
     }
     return render(request, 'update.html', context)
 
+
 def restaurant_delete(request, restaurant_id):
+    if not request.user.is_staff:
+        return redirect("no-access")
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+def no_access(request):
+    return render(request, "noaccess.html")
